Fixing line length, refactoring file.
